requirements.txt
SQLAlchemy==2.0.40
et_xmlfile==2.0.0
numpy==2.2.4
openpyxl==3.1.5
pandas==2.2.3
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
tzdata==2025.2
python-dotenv==1.1.0